pyproject.toml
[project]
name = "capability-runtime"
version = "0.0.0"
description = "Bridge/glue layer that composes Agently (LLM/TriggerFlow) with skills-runtime-sdk (skills/tools/WAL/events)."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
dependencies = [
  "PyYAML>=6",
  "pydantic>=2,<3",
  "agently==4.0.8",
  "skills-runtime-sdk==0.1.6",
]

[project.optional-dependencies]
dev = [
  "pytest>=7",
  "pytest-asyncio>=0.23",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
addopts = "-q"
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
  "integration: optional integration tests (may require docker/services)",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
